file renaming to match class renaming
(cherry picked from commit 7b2b5645c6ef96206b417bdbbc0b19ea54517626)
