+++ pacman/model/graphs/__init__.py
@@ -14,12 +14,12 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from .abstract_edge import AbstractEdge
-from .abstract_fpga_vertex import AbstractFPGA
+from .abstract_fpga import AbstractFPGA
 from .abstract_graph import AbstractGraph
 from .abstract_outgoing_edge_partition import AbstractOutgoingEdgePartition
-from .abstract_spinnaker_link_vertex import AbstractSpiNNakerLink
+from .abstract_spinnaker_link import AbstractSpiNNakerLink
 from .abstract_vertex import AbstractVertex
-from .abstract_virtual_vertex import AbstractVirtual
+from .abstract_virtual import AbstractVirtual
 
 __all__ = ["AbstractEdge", "AbstractFPGA", "AbstractGraph",
            "AbstractOutgoingEdgePartition", "AbstractSpiNNakerLink",
