+++ pacman/model/graphs/abstract_fpga.py
@@ -14,7 +14,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from spinn_utilities.abstract_base import abstractproperty
-from .abstract_virtual_vertex import AbstractVirtual
+from .abstract_virtual import AbstractVirtual
 
 
 class AbstractFPGA(AbstractVirtual):
