+++ pacman/model/graphs/abstract_virtual.py
