+++ pacman/operations/partition_algorithms/partition_and_place_partitioner.py
@@ -18,7 +18,7 @@ import logging
 from six import raise_from
 from spinn_utilities.progress_bar import ProgressBar
 from pacman.exceptions import PacmanPartitionException, PacmanValueError
-from pacman.model.graphs.abstract_virtual_vertex import AbstractVirtual
+from pacman.model.graphs.abstract_virtual import AbstractVirtual
 from pacman.model.constraints.partitioner_constraints import (
     AbstractPartitionerConstraint, MaxVertexAtomsConstraint,
     FixedVertexAtomsConstraint, SameAtomsAsVertexConstraint)
